split handler/services per cloud, assign ad group lambda
