+++ server/config-description.json
@@ -10,5 +10,7 @@
   "adRunbookUrl": "URL of Runbook which creates AD group and assigns owner. This group gets propagated to AWS SSO",
   "adRunbookKey": "API key for Runbook",
   "gcpAccountKeys": "JSON for Service account access, it contains privete key for auth and other service values",
-  "gcpParentFolderValue": "Value of gcp parent project"
+  "gcpParentFolderValue": "Value of gcp parent project",
+  "gcpOrganization": "Name of gcp organization"
+
 }
