+++ server/serverless.yml
@@ -45,6 +45,7 @@ provider:
     SNS_ALERTS_TOPIC: arn:aws:sns:${self:provider.region}:#{AWS::AccountId}:sky9-${self:provider.stage}-alerts-topic
     GCP_SERVICE_ACCOUNT: ${ssm:/sky9/${self:provider.stage}/gcpAccountKeys}
     GCP_PARENT_FOLDER_VALUE: ${ssm:/sky9/${self:provider.stage}/gcpParentFolderValue, "320110726660"}
+    GCP_ORGANIZATION: ${ssm:/sky9/${self:provider.stage}/gcpOrganization, "GCP-SRB"}
 
 plugins:
   - serverless-pseudo-parameters
@@ -54,12 +55,12 @@ plugins:
   - serverless-dashboard
 
 functions:
-  - ${file(./serverless-account-create.yml):functions}
-  - ${file(./serverless-account-crud.yml):functions}
+  - ${file(./serverless-aws-account-create.yml):functions}
+  - ${file(./serverless-aws-account-crud.yml):functions}
   - ${file(./serverless-gcp-account-create.yml):functions}
 
 stepFunctions:
-  - ${file(./serverless-account-create.yml):stepFunctions}
+  - ${file(./serverless-aws-account-create.yml):stepFunctions}
   - ${file(./serverless-gcp-account-create.yml):stepFunctions}
 
 custom:
