+++ server/sls-env.yml
@@ -14,6 +14,7 @@ default: &default
   SNS_ALERTS_TOPIC: arn:aws:sns:${self:provider.region}:#{AWS::AccountId}:sky9-${self:provider.stage}-alerts-topic
   GCP_SERVICE_ACCOUNT: ${ssm:/sky9/${self:provider.stage}/gcpAccountKeys}
   GCP_PARENT_FOLDER_VALUE: ${ssm:/sky9/${self:provider.stage}/gcpParentFolderValue}
+  GCP_ORGANIZATION: ${ssm:/sky9/${self:provider.stage}/gcpOrganization}
 
 local:
   <<: *default
