+++ server/src/utils.js
@@ -1,0 +1,4 @@
+export const clouds = {
+  AWS: 'AWS',
+  GCP: 'GCP'
+}
